Adds an <F4> command to open the current HTML file in the default browser.
Adds functions to generate HTML tags and generate abbreviation for HTML
tags using a list. The abbreviations start with the letter 't' or the
letters 'it' for an inline tag followed by the 'name' of the tag.
Also the 'tag' and 'itag' abbreviations let you input a general tag
(inline or not) by supplying its name in the input generated by the
function.
In the current implementation the space inserted after the abbreviation
is not removed as it should.