FIXE INDENTATION DISPLACEMENT
Fixe indentation displacement when commenting with a comment symbol with
more than one character. Also remove the 'very-magic' option in regular
expresions to have a simplers list of special character for comment
symbols.

The commenting of a line does not change the indentation if the
indentation is greater than the length of the comment symbol.
Otherwise, the indentation may not be respected.
example :

        Indented line to comment
//      Indented line commented
        Indented line uncommented

Non-indented line to comment
//Non-indented line commented
Non-indented line uncommented

 Indented line to comment
//Indented line commented
Changed indentation of the uncommented line