EDIT FILE UPON ENTERING BUFFER
Upon editing a file named 'file.ext' and saving it to a new name
'newfile.ext', the header of the newly created file will correspond to
the one of 'file.ext'. To fix this, we edit files upon entering them and
verify that the commented name of the file correspond to the name of the
actual file.

Note that it would be better to make those changes upon the creation of
the new file, but changes take effect in the current buffer
('file.ext').