EXTENDS GENERATEDOCUMENTTYPE PLUGIN
GenerateDocumentType is a plugin that creates XML DocumentType
from a given set of data such as ./tests/test.txt :

  #depth    tagName   tagType   mandatory standaloneName
  0         tag0      document  true      tag0
  1         tag1      document  true      tag1
  2         tag2      string    true
  2         tag3      document  true      tag3
  3         tag4      string    false
  1         tag5      document  false     tag5
  2         tag6      string    true
  0         tag7      string    true
  0         tag8      string    true

From such a data set, GenerateDocumentType will create an XML
file for each root node associated with a standaloneName.
Such a file will represent a root node and all of its children and
siblings. When a standalone node is the child of a node, a reference to its
standalone file will be inserted (to implemented yet).

The depth represents the depth of the element.

The tagName represents the name of the element.

The tagType represent the type of the element represented by that node
in the Document Type.

The mandatory attribute specify whether that element is mandatory.

The standaloneName represent the name of the file for the standalone
element.

GENERATEDOCUMENTTYPE includes a function (not mapped yet) to translate
the graphical depth of an element to a numerical value.

  by example :
    the following :
      depth0
      --> depth1
      ----> depth2
    is replaced by :
      0 depth0
      1 depth1
      2 depth2